field and robot estimation is working

+++ simulation/TrueBattleBotSim/Assets/RosMessages/BwInterfaces/msg/UVKeypointMsg.cs
@@ -13,16 +13,16 @@ namespace RosMessageTypes.BwInterfaces
         public const string k_RosMessageName = "bw_interfaces/UVKeypoint";
         public override string RosMessageName => k_RosMessageName;
 
-        public float x;
-        public float y;
+        public uint x;
+        public uint y;
 
         public UVKeypointMsg()
         {
-            this.x = 0.0f;
-            this.y = 0.0f;
+            this.x = 0;
+            this.y = 0;
         }
 
-        public UVKeypointMsg(float x, float y)
+        public UVKeypointMsg(uint x, uint y)
         {
             this.x = x;
             this.y = y;
